requirements.txt
gradio>=3.50.0
groq>=0.4.0
pillow>=9.0.0
requests>=2.28.0
python-dotenv>=0.20.0
torch
diffusers
transformers 
